feat(starknet_consensus_orchestrator): move ContextConfig to orchestrator crate
