+++ crates/papyrus_node/Cargo.toml
@@ -39,6 +39,7 @@ starknet_api.workspace = true
 starknet_class_manager_types.workspace = true
 starknet_client.workspace = true
 starknet_consensus.workspace = true
+starknet_consensus_orchestrator.workspace = true
 strum.workspace = true
 tokio = { workspace = true, features = ["full", "sync"] }
 tracing.workspace = true
