+++ crates/starknet_consensus/src/types.rs
@@ -1,13 +1,9 @@
 //! Types for interfacing between consensus and the node.
-
-use std::collections::BTreeMap;
 use std::fmt::Debug;
 use std::time::Duration;
 
 use async_trait::async_trait;
 use futures::channel::{mpsc, oneshot};
-use papyrus_config::dumping::{ser_param, SerializeConfig};
-use papyrus_config::{ParamPath, ParamPrivacyInput, SerializedParam};
 use papyrus_network::network_manager::{
     BroadcastTopicChannels,
     BroadcastTopicClient,
@@ -16,10 +12,8 @@ use papyrus_network::network_manager::{
 use papyrus_network_types::network_types::BroadcastedMessageMetadata;
 use papyrus_protobuf::consensus::{ProposalFin, ProposalInit, Vote};
 use papyrus_protobuf::converters::ProtobufConversionError;
-use serde::{Deserialize, Serialize};
 use starknet_api::block::{BlockHash, BlockNumber};
-use starknet_api::core::{ChainId, ContractAddress};
-use validator::Validate;
+use starknet_api::core::ContractAddress;
 
 /// Used to identify the node by consensus.
 /// 1. This ID is derived from the id registered with Starknet's L2 staking contract.
@@ -29,72 +23,6 @@ use validator::Validate;
 pub type ValidatorId = ContractAddress;
 pub type Round = u32;
 pub type ProposalCommitment = BlockHash;
-
-// TODO(guyn): move this to another file.
-/// Configuration for the Context struct.
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Validate)]
-pub struct ContextConfig {
-    /// The buffer size for streaming outbound proposals.
-    pub proposal_buffer_size: usize,
-    /// The number of validators.
-    pub num_validators: u64,
-    /// The chain id of the Starknet chain.
-    pub chain_id: ChainId,
-    /// Maximum allowed deviation (seconds) of a proposed block's timestamp from the current time.
-    pub block_timestamp_window: u64,
-    /// The data availability mode, true: Blob, false: Calldata.
-    pub l1_da_mode: bool,
-}
-
-impl SerializeConfig for ContextConfig {
-    fn dump(&self) -> BTreeMap<ParamPath, SerializedParam> {
-        BTreeMap::from_iter([
-            ser_param(
-                "proposal_buffer_size",
-                &self.proposal_buffer_size,
-                "The buffer size for streaming outbound proposals.",
-                ParamPrivacyInput::Public,
-            ),
-            ser_param(
-                "num_validators",
-                &self.num_validators,
-                "The number of validators.",
-                ParamPrivacyInput::Public,
-            ),
-            ser_param(
-                "chain_id",
-                &self.chain_id,
-                "The chain id of the Starknet chain.",
-                ParamPrivacyInput::Public,
-            ),
-            ser_param(
-                "block_timestamp_window",
-                &self.block_timestamp_window,
-                "Maximum allowed deviation (seconds) of a proposed block's timestamp from the \
-                 current time.",
-                ParamPrivacyInput::Public,
-            ),
-            ser_param(
-                "l1_da_mode",
-                &self.l1_da_mode,
-                "The data availability mode, true: Blob, false: Calldata.",
-                ParamPrivacyInput::Public,
-            ),
-        ])
-    }
-}
-
-impl Default for ContextConfig {
-    fn default() -> Self {
-        Self {
-            proposal_buffer_size: 100,
-            num_validators: 1,
-            chain_id: ChainId::Mainnet,
-            block_timestamp_window: 1,
-            l1_da_mode: true,
-        }
-    }
-}
 
 /// Interface for consensus to call out to the node.
 ///
