+++ crates/starknet_consensus_orchestrator/Cargo.toml
@@ -33,6 +33,7 @@ tokio = { workspace = true, features = ["full"] }
 tokio-util = { workspace = true, features = ["rt"] }
 tracing.workspace = true
 url = { workspace = true, features = ["serde"] }
+validator.workspace = true
 
 [dev-dependencies]
 cairo-lang-casm.workspace = true
