+++ crates/starknet_consensus_orchestrator/src/lib.rs
@@ -14,3 +14,6 @@ pub mod fee_market;
 
 /// Consensus' versioned constants.
 pub mod orchestrator_versioned_constants;
+
+/// The orchestrator's configuration.
+pub mod config;
